102-fibonacci.c

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -1,16 +1,16 @@
 #include <stdio.h>
 
 void print_fibonacci_sequence(int n) {
-    unsigned long long first = 1;
-    unsigned long long second = 2;
-    unsigned long long next;
+    int first = 1;
+    int second = 2;
+    int next;
     int i;
 
-    printf("%llu, %llu", first, second);
+    printf("%d, %d", first, second);
 
     for (i = 3; i <= n; i++) {
         next = first + second;
-        printf(", %llu", next);
+        printf(", %d", next);
         first = second;
         second = next;
     }
